+++ src/sn/FocusMng.ts
@@ -1,5 +1,5 @@
 /* ***** BEGIN LICENSE BLOCK *****
-	Copyright (c) 2020-2024 Famibee (famibee.blog38.fc2.com)
+	Copyright (c) 2020-2025 Famibee (famibee.blog38.fc2.com)
 
 	This software is released under the MIT License.
 	http://opensource.org/licenses/mit-license.php